test-driver integration changes
