+++ src/events/IonEventStream.ts
@@ -13,36 +13,29 @@
  * permissions and limitations under the License.
  */
 
-import {Reader} from "./IonReader";
+import {Reader} from "../IonReader";
 import {IonEvent, IonEventFactory, IonEventType} from "./IonEvent";
-import {Writer} from "./IonWriter";
-import {IonType} from "./IonType";
-import {IonTypes} from "./IonTypes";
-import {defaultLocalSymbolTable, makeReader, ReaderOctetBuffer} from "./Ion";
-import {BinaryReader} from "./IonBinaryReader";
-import {BinarySpan} from "./IonSpan";
-import {
-    ComparisonResult,
-    ComparisonResultType,
-    ErrorType,
-    IonCliCommonArgs,
-    IonCliError,
-    IonComparisonReport
-} from "./Cli";
-import {ComparisonType} from "./Compare";
+import {Writer} from "../IonWriter";
+import {IonType} from "../IonType";
+import {IonTypes} from "../IonTypes";
+import {defaultLocalSymbolTable, makeReader, ReaderOctetBuffer} from "../Ion";
+import {BinaryReader} from "../IonBinaryReader";
+import {BinarySpan} from "../IonSpan";
+import {ComparisonResultType} from "../ComparisonResult";
+import {ComparisonResult} from "../ComparisonResult";
 
 export class IonEventStream {
-    private eventStream: IonEvent[];
+    eventStream: IonEvent[];
     private reader: Reader;
     private eventFactory: IonEventFactory;
     isEventStream: boolean; // whether the reader has an event stream as input
 
-    constructor(reader: Reader, path: string = "", args?: IonCliCommonArgs) {
+    constructor(reader: Reader) {
         this.eventStream = [];
         this.reader = reader;
         this.eventFactory = new IonEventFactory();
         this.isEventStream = false;
-        this.generateStream(path, args!);
+        this.generateStream();
     }
 
     writeEventStream(writer: Writer) {
@@ -140,10 +133,8 @@ export class IonEventStream {
 
     /**
      *  compares eventstreams and generates comparison result
-     *
-     *  @param comparisonReport: optional argument to write a comparison report for the equivalence result
      */
-    compare(expected: IonEventStream, comparisonReport?: IonComparisonReport): ComparisonResult {
+    compare(expected: IonEventStream): ComparisonResult {
         let actualIndex: number = 0;
         let expectedIndex: number = 0;
         if(this.eventStream.length != expected.eventStream.length) {
@@ -159,8 +150,8 @@ export class IonEventStream {
                 case IonEventType.SCALAR: {
                     let eventResult = actualEvent.compare(expectedEvent);
                     if (eventResult.result == ComparisonResultType.NOT_EQUAL) {
-                        if(comparisonReport)
-                            comparisonReport.writeComparisonReport(ComparisonResultType.NOT_EQUAL, eventResult.message, actualIndex, expectedIndex);
+                        eventResult.actualIndex = actualIndex;
+                        eventResult.expectedIndex = expectedIndex;
                         return eventResult;
                     }
                     break;
@@ -170,8 +161,9 @@ export class IonEventStream {
                     if (eventResult.result == ComparisonResultType.NOT_EQUAL) {
                         actualIndex += eventResult.actualIndex;
                         expectedIndex += eventResult.expectedIndex;
-                        if(comparisonReport)
-                            comparisonReport.writeComparisonReport(ComparisonResultType.NOT_EQUAL, eventResult.message, actualIndex, expectedIndex);
+
+                        eventResult.actualIndex = actualIndex;
+                        eventResult.expectedIndex = expectedIndex;
                         return eventResult;
                     } else {
                         actualIndex = actualIndex + actualEvent.ionValue.length;
@@ -185,27 +177,13 @@ export class IonEventStream {
                     break;
                 }
                 default: {
-                    let message = "Unexpected event type: " + actualEvent.eventType;
-                    if(comparisonReport) {
-                        comparisonReport.writeComparisonReport(ComparisonResultType.ERROR,
-                            message, actualIndex, expectedIndex);
-                    } else {
-                        throw new Error(message);
-                    }
+                    throw new Error("Unexpected event type: " + actualEvent.eventType);
                 }
             }
             actualIndex++;
             expectedIndex++;
         }
         return new ComparisonResult(ComparisonResultType.EQUAL);
-    }
-
-    /**
-     * Returns true if this `IonEventStream` is equal to the provided `IonEventStream`
-     * as determined by the `ComparisonType`; returns false otherwise.
-     * */
-    equivs(expected: IonEventStream, comparisonType: ComparisonType): boolean {
-        return this.compareEquivs(expected, comparisonType).result == ComparisonResultType.EQUAL;
     }
 
     isEmbedded(event: IonEvent): boolean {
@@ -215,107 +193,7 @@ export class IonEventStream {
         return false;
     }
 
-    /**
-     *  equivs, non-equivs & equiv-timeline comparison of eventstreams
-     *
-     *  @param comparisonReport: optional argument to write a comparison report for the equivalence result
-     */
-    compareEquivs(expected: IonEventStream , comparisonType: ComparisonType, comparisonReport?: IonComparisonReport): ComparisonResult {
-        let actualIndex: number = 0;
-        let expectedIndex: number = 0;
-
-        while (actualIndex < this.eventStream.length && expectedIndex < expected.eventStream.length) {
-            let actualEvent = this.eventStream[actualIndex];
-            let expectedEvent = expected.eventStream[expectedIndex];
-
-            if(actualEvent.eventType == IonEventType.STREAM_END && expectedEvent.eventType == IonEventType.STREAM_END) {
-                break;
-            } else if(actualEvent.eventType == IonEventType.STREAM_END || expectedEvent.eventType == IonEventType.STREAM_END) {
-                throw new Error("Different number of comparison sets.");
-            } else if (!(actualEvent.ionType == IonTypes.LIST || actualEvent.ionType == IonTypes.SEXP)
-                || !(expectedEvent.ionType == IonTypes.LIST || expectedEvent.ionType == IonTypes.SEXP)) {
-                throw new Error("Comparison sets must be lists or s-expressions.");
-            } else if(this.isEmbedded(actualEvent) as any ^ this.isEmbedded(expectedEvent) as any) {
-                throw new Error("Both streams should be embedded streams.");
-            }
-
-            // both containers has type any as it maybe IonEventStream[] or IonEvent[]
-            // depending on the type compare method will be called
-            let actualContainer: any = [];
-            let expectedContainer: any = [];
-            if (this.isEmbedded(actualEvent) && this.isEmbedded(expectedEvent)) {
-                //we found a list of strings that we need to interpret as top level ion text streams.
-                actualContainer = this.parseEventStream(actualEvent);
-                expectedContainer = this.parseEventStream(expectedEvent);
-            } else {//we're in an sexp/list
-                actualContainer = this.parseContainer(actualEvent);
-                expectedContainer = this.parseContainer(expectedEvent);
-            }
-
-            actualIndex = actualIndex + actualEvent.ionValue.length;
-            expectedIndex = expectedIndex + expectedEvent.ionValue.length;
-
-            for (let i = 0; i < actualContainer.length; i++) {
-                for (let j = 0; j < expectedContainer.length; j++) {
-                    // for non-equivs: not comparing same index's value as it will always be same.
-                    if (comparisonType == ComparisonType.NON_EQUIVS && i == j)
-                        continue;
-                    let actualContainerEvent: IonEvent = actualContainer[i];
-                    let expectedContainerEvent: IonEvent = expectedContainer[j];
-                    let eventResult;
-
-                    if (comparisonType == ComparisonType.EQUIV_TIMELINE && actualContainerEvent.ionType == IonTypes.TIMESTAMP) {
-                        let ionTimestampActual = actualContainerEvent.ionValue;
-                        let ionTimestampExpected = expectedContainerEvent.ionValue;
-                        eventResult = ionTimestampActual.compareTo(ionTimestampExpected) == 0 ? new ComparisonResult(ComparisonResultType.EQUAL)
-                            : new ComparisonResult(ComparisonResultType.NOT_EQUAL, ionTimestampActual + " vs. " + ionTimestampExpected);
-                    } else {
-                        eventResult = actualContainerEvent.compare(expectedContainerEvent);
-                    }
-
-                    if ((comparisonType == ComparisonType.EQUIVS || comparisonType == ComparisonType.EQUIV_TIMELINE)
-                        && eventResult.result == ComparisonResultType.NOT_EQUAL) {
-                        if(comparisonReport) {
-                            comparisonReport.writeComparisonReport(ComparisonResultType.NOT_EQUAL, eventResult.message, i + 1, j + 1);
-                        }
-                        return new ComparisonResult(ComparisonResultType.NOT_EQUAL);
-                    } else if (comparisonType == ComparisonType.NON_EQUIVS && eventResult.result == ComparisonResultType.EQUAL) {
-                        if(comparisonReport) {
-                            comparisonReport.writeComparisonReport(ComparisonResultType.EQUAL,
-                                "Both values are equal in non-equivs comparison.", i + 1, j + 1);
-                        }
-                        return new ComparisonResult(ComparisonResultType.EQUAL);
-                    }
-                }
-            }
-            actualIndex++;
-            expectedIndex++;
-        }
-        return new ComparisonResult(comparisonType == ComparisonType.NON_EQUIVS ? ComparisonResultType.NOT_EQUAL : ComparisonResultType.EQUAL);
-    }
-
-    private parseContainer(event: IonEvent): IonEvent[] {
-        let container: IonEvent[] = [];
-        for (let j = 0; j < event.ionValue.length - 1; j++) {
-            container.push(event.ionValue[j]);
-            if (event.ionValue[j].eventType === IonEventType.CONTAINER_START) {
-                j += event.ionValue[j].ionValue.length;
-            }
-        }
-        return container;
-    }
-
-    private parseEventStream(event: IonEvent): any {
-        let container: any = [];
-        for (let j = 0; j < event.ionValue.length - 1; j++) {
-            if(event.ionValue[j].eventType == IonEventType.STREAM_END)
-                continue;
-            container.push(new IonEventStream(makeReader(event.ionValue[j].ionValue)));
-        }
-        return container;
-    }
-
-    private generateStream(path: string, args: IonCliCommonArgs): void {
+    private generateStream(): void {
         try{
             let tid: IonType | null = this.reader.next();
             if (tid === IonTypes.SYMBOL && this.reader.stringValue() === "$ion_event_stream") {
@@ -358,10 +236,8 @@ export class IonEventStream {
                 }
                 tid = this.reader.next();
             }
-        } catch (Error) {
-            if(args) {
-                new IonCliError(ErrorType.READ, path, Error.message, args.getErrorReportFile(), this.eventStream.length).writeErrorReport();
-            }
+        } catch (er) {
+            throw new Error("READ " + er.message);
         }
     }
