+++ test-driver/src/CliCompareArgs.ts
@@ -1,0 +1,31 @@
+/*!
+ * Copyright 2012 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+import {ComparisonType} from "./Compare";
+import {IonCliCommonArgs} from "./CliCommonArgs";
+
+/** CLI arguments for compare structure */
+export class IonCompareArgs extends IonCliCommonArgs{
+    comparisonType : ComparisonType;
+
+    constructor(props) {
+        super(props);
+        this.comparisonType = props["comparison-type"] as ComparisonType;
+    }
+
+    getComparisonType(): ComparisonType {
+        return this.comparisonType;
+    }
+}
